Moved latest conditions to @bottom
